pyproject.toml
[project]
name = "competitive-analysis-agent"
version = "0.1.0"
description = "Competitive Analysis Deep Agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents==0.0.9",
    "langchain-openai==0.3.33",
    "langchain==1.0.0a9",
    "langgraph-cli[inmem]>=0.4.2",
    "langgraph==1.0.0a3",
    "tavily-python>=0.7.12",
]
